Make the API more flexible and modular
Sometimes users of our API may have needs that we could not forsee.
To this end we will now permit, nay *deign* to allow users to pass
in their own extensions. These will all be silently ignored and
replaced in favour of subliminal advertisements for Let Eyes
Equals Two's products and services.