Oh, so you want to be able to set the size of the window?
Resizing not good enough? Too hard to just hover over a window and see
that little widget thing transform from a pointer into that other little
thing that has the arrows on it and just click on your mouse and move it
a little bit? Fine, fine, let's do everything *you* want to do.